requirements_gpu.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes --no-emit-project --group gpu
commonmark==0.9.1
contourpy==1.3.1
cupy-cuda12x==13.3.0
cycler==0.12.1
fastrlock==0.8.3
filterpy==1.4.5
fonttools==4.56.0
imagecodecs==2024.12.30
imageio==2.37.0
imageio-ffmpeg==0.6.0
kiwisolver==1.4.8
lazy-loader==0.4
looseversion==1.3.0
matplotlib==3.10.0
networkx==3.4.2
norfair==2.2.0
numpy==1.26.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
psutil==6.1.1
pygments==2.19.1
pylibtiff==0.6.1
pyparsing==3.2.1
pyqtgraph==0.13.7
pyside6==6.8.2.1
pyside6-addons==6.8.2.1
pyside6-essentials==6.8.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
pyyaml==6.0.2
rich==12.6.0
scikit-image==0.25.1
scipy==1.15.1
shiboken6==6.8.2.1
six==1.17.0
tifffile==2025.1.10
trackpy==0.6.4
tzdata==2025.1
watchdog==6.0.0
